# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sinema API",
        "description": "Film, tür, kişi ve oyuncu yönetimi için RESTful API",
        "contact": {
            "name": "API Geliştirici",
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/turler": {
            "get": {
                "tags": [
                    "Türler"
                ],
                "summary": "Tüm türleri listele",
                "description": "Veritabanındaki tüm film türlerini getirir",
                "operationId": "2d9cef559cf62d53e6a13f8abd1eefc5",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tur"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Türler"
                ],
                "summary": "Yeni tür ekle",
                "description": "Veritabanına yeni bir film türü ekler",
                "operationId": "68820404490b9855d520a6fae79a7165",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Tür adı",
                                        "type": "string",
                                        "example": "Aksiyon"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tür başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Tur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            }
        },
        "/turler/{id}": {
            "get": {
                "tags": [
                    "Türler"
                ],
                "summary": "Belirli türü getir",
                "description": "ID'ye göre belirli bir film türünü getirir",
                "operationId": "cc98978ea99019c327059eb15f2c73f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tür ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Tur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tür bulunamadı"
                    }
                }
            },
            "put": {
                "tags": [
                    "Türler"
                ],
                "summary": "Tür güncelle",
                "description": "Belirli bir film türünün bilgilerini günceller",
                "operationId": "23194461677df1f2d5694f1b5cb65287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tür ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Tür adı",
                                        "type": "string",
                                        "example": "Aksiyon"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tür başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Tur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tür bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Türler"
                ],
                "summary": "Tür sil",
                "description": "Belirli bir film türünü veritabanından siler",
                "operationId": "8d91166378d45d438701f02c05fc181c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tür ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tür başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tür başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tür bulunamadı"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tur": {
                "title": "Tür",
                "description": "Film türü modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Tür adı",
                        "type": "string",
                        "example": "Aksiyon"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}